third_party/WebKit/LayoutTests/fast/text/first-letter-bad-line-boxes-crash-expected.txt

CONSOLE WARNING: We don't execute document.execCommand() this time, because it is called recursively.
CONSOLE WARNING: We don't execute document.execCommand() this time, because it is called recursively.
CONSOLE ERROR: line 66: Uncaught TypeError: Failed to execute 'surroundContents' on 'Range': parameter 1 is not of type 'Node'.
\\\\\\m{VVVVVVVV\\\\\\m{VVVVVVVVVVVVVVV<};<WWWWW
 \\\\\\m{VVVVVVVVVVVVVVV<};<WWWWW
xt
